// src/types/adsTypes/adItem.ts
export interface AdItem {
    date: string;
    campaignId: string;
    adGroupId: string;
    itemId: string;
    numAdsShown: string;
    numAdsClicks: string;
    adSpend: string;
    attributedOrders14days: string;
    attributedSales14days: string;
    advertisedSkuSales14days: string;
    otherSkuSales14days: string;
    unitsSold14days: string;
    advertisedSkuUnits14days: string;
    otherSkuUnits14days: string;
    searchedKeyword: string;
    store_id: number;
    itemName: string;
    matchType: string;
}